Phase 4a: effects - per-voice lowpass, fundsp reverb, delay sends
- rudel-dsp: per-voice RBJ biquad low-pass (cutoff/resonance), plus room and
  delay send amounts on VoiceParams.
- rudel-audio: global stereo feedback delay and a fundsp FDN reverb_stereo on
  send buses, mixed per frame in the audio callback.
- Widen Frac from Rational64 to Ratio<i128> and quantize Frac::from_f64 to a
  bounded grid. Fixes a scheduler-thread "multiply with overflow" panic when
  querying at arbitrary f64 cycle positions (the overflow risk flagged in the
  plan's open items).

Effects demo updated (filter sweep + reverb + delay). 59 workspace tests pass
(incl. filter attenuation, delay echo, reverb tail); clippy clean.

+++ crates/rudel-core/src/pattern.rs
@@ -557,7 +557,7 @@ pub fn slowcat(pats: &[Pattern]) -> Pattern {
     let len = pats.len() as i64;
     Pattern::new(move |state| {
         let span = state.span;
-        let pat_n = span.begin.sam().numer().rem_euclid(len) as usize;
+        let pat_n = span.begin.sam().numer().rem_euclid(len as i128) as usize;
         let pat = &pats[pat_n];
         // Keep cycles from constituent patterns from being skipped.
         let offset = span.begin.floor() - (span.begin / Frac::int(len)).floor();
@@ -581,7 +581,7 @@ pub fn slowcat_prime(pats: &[Pattern]) -> Pattern {
         if len == 0 {
             return vec![];
         }
-        let pat_n = state.span.begin.sam().numer().rem_euclid(len) as usize;
+        let pat_n = state.span.begin.sam().numer().rem_euclid(len as i128) as usize;
         pats[pat_n].query(state)
     })
     .split_queries()
